fix(frontend): cleanup and improved functionality
